refactor code to support multiple consumers and producers

+++ pkg/consumer/consumer.go
@@ -2,6 +2,6 @@ package consumer
 
 type ConsumerProvider interface {
 	Get(string) interface{}
-	GetKeys(string, int64) []string
+	GetKeys(string, int) []string
 	DeleteKey(string)
 }
